requirements.txt
torch>=2.0.0
clip @ git+https://github.com/openai/CLIP.git
opencv-python>=4.5.0
ffmpeg-python>=0.2.0
yt-dlp>=2023.0.0
numpy>=1.20.0 